Update package name/application id to: io.homeassistant.companion.android Update version code and name to be equal to that of the regular android app Merge settings.gradle.kts include into one line Move Home.kt to new path (based on package name)
